AbstractDirectoryServiceをインターフェースに変更し、CopyDirectoryService、MoveDirectoryService、DiffDirectoryServiceを実装クラスに更新
