+++ src/test/java/kmg/tool/directorytool/service/AbstractDirectoryServiceTest.java
@@ -12,6 +12,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 import org.springframework.boot.test.context.SpringBootTest;
 
+import kmg.tool.directorytool.service.impl.AbstractDirectoryServiceImpl;
+
 /**
  * ディレクトリ操作サービスのテストのための基底クラス。 共通のセットアップ、検証、クリーンアップ機能を提供する。
  */
@@ -160,7 +162,7 @@ public abstract class AbstractDirectoryServiceTest {
         Files.writeString(file1, "content1");
         Files.writeString(file2, "different content");
 
-        Assertions.assertFalse(AbstractDirectoryService.compareFiles(file1, file2), "異なるサイズのファイルは不一致となるべき");
+        Assertions.assertFalse(AbstractDirectoryServiceImpl.compareFiles(file1, file2), "異なるサイズのファイルは不一致となるべき");
 
     }
 
@@ -181,7 +183,7 @@ public abstract class AbstractDirectoryServiceTest {
         Files.writeString(sourceFile, "test content");
 
         // サービスをオーバーライドして例外をシミュレート
-        final AbstractDirectoryService errorService = new AbstractDirectoryService() {
+        final AbstractDirectoryService errorService = new AbstractDirectoryServiceImpl() {
 
             @Override
             protected void processPath(final Path sourcePath, final Path targetPath, final Path relativePath)
@@ -222,7 +224,7 @@ public abstract class AbstractDirectoryServiceTest {
         Files.writeString(sourceFile, "test content");
 
         // サービスをオーバーライドして長時間のタスクをシミュレート
-        final AbstractDirectoryService slowService = new AbstractDirectoryService() {
+        final AbstractDirectoryService slowService = new AbstractDirectoryServiceImpl() {
 
             @Override
             protected void processPath(final Path sourcePath, final Path targetPath, final Path relativePath)
